build: create a specific browser bundle
