+++ src/browser.js
@@ -1,0 +1,7 @@
+'use strict'
+
+module.exports = require('./factory')({
+  mql: require('@microlink/mql'),
+  toCompress: code => code.toString(),
+  VERSION: '__VERSION__'
+})
